% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim_timespan.R
\name{anim_timespan}
\alias{anim_timespan}
\title{Get dates of individuals first and last sample}
\usage{
anim_timespan(individual_id, sample_date, mortality_sample)
}
\arguments{
\item{individual_id}{Column in the dataframe of all samples containing
individual animal identifier code.
Defined as \code{dataframe$column}.}

\item{sample_date}{Column in the dataframe of all samples containing
the date of sample collection.
Must be in \code{Date} format. Defined as \code{dataframe$column}.}

\item{mortality_sample}{Logical vector or column in the dataframe of all samples that
identifies samples that represent a mortality event (e.g. from a dead animal).
\code{TRUE} values mark mortality samples.
Defined as \code{dataframe$column} or a standalone logical vector the same length
as \code{individual_id}.}
}
\value{
A data frame with four columns and one row for each \code{individual_id}.
Returned data frame columns correspond to individual identification key (\code{ID}),
date of first (\code{FirstSeen}) and last (\code{LastSeen}) sample of individual and
logical (\code{TRUE/FALSE}) value that identifies if the individual is dead (\code{IsDead}).
}
\description{
Takes data frame of all samples and returns the dates of individuals first
and last sample.
Besides that the function determines if an individual is dead based on the
mortality_sample column.
}
\examples{
anim_timespan(
  individual_id = wolf_samples$AnimalRef,
  sample_date = wolf_samples$Date,
  mortality_sample = wolf_samples$IsMortality
)
}
