% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_table.R
\name{fam_table}
\alias{fam_table}
\title{Prepares pedigree data for plotting and spatial representation}
\usage{
fam_table(
  fams,
  all.fams,
  ped,
  sampledata,
  datacolumns = c("Sample", "AnimalRef", "GeneticSex", "Date", "SType", "X", "Y",
    "FirstSeen", "LastSeen", "IsDead"),
  deadSample = c("Tissue", "Decomposing Tissue", "Bone")
)
}
\arguments{
\item{fams}{data frame. (Subset of) family (\code{fams}) data generated by \code{\link{org_fams}} function.
Just families that we want to plot have to be selected.}

\item{all.fams}{data frame. Family (\code{fams}) data generated by \code{\link{org_fams}} function.}

\item{ped}{data frame. Organized pedigree (\code{ped}) generated by \code{\link{org_fams}} function.}

\item{sampledata}{data frame. Metadata for all genetic samples that belong
to the individuals included in pedigree reconstruction analysis.
For description of \code{sampledata} structure and sample information needed for \code{fam_table()} see Details.}

\item{datacolumns}{vector of column names included \code{sampledata} that are needed to produce
this functions output (see Details).}

\item{deadSample}{single value or vector of different lethal sample types.
Defaults to c("Tissue", "Decomposing Tissue", "Bone").}
}
\value{
Extended \code{sampledata} data frame that includes all columns defined in \code{datacolumns}
parameter and adds information needed for visual and spatial representation of
pedigree:
\itemize{
\item \code{plottingID}: numeric. Identifier number for temporal pedigree plot
\code{\link{ped_satplot}}. In case of polygamous animals same individual
can be included in more than one family,
\item \code{FamID}: numeric. Identifier number of family that individual belongs to,
\item \code{polyCluster}: numeric. Identifier number for the poygami cluster of individual,
\item \code{rep}: logical. Is individual reproductive in current family,
(current family defined with \code{FamID} for a particular entry),
\item \code{later_rep}: logical. Is individual reproductive in any other (later) families,
\item \code{isPolygamous}: logical. Has individual more than one mate,
\item \code{dead}: logical. Is individual dead,
\item \code{first_sample}: logical. Is this particular sample the first sample of individual,
\item \code{last_sample}: logical. Is this particular sample the last sample of individual,
\item \code{isReference}: logical. Is this particular sample reference sample of individual.
}
}
\description{
\code{fam_table} combines pedigree and sample data for visual (\code{\link{ped_satplot}})
and spatial (\code{\link{ped_spatial}}) representation of the pedigree.
}
\details{
\itemize{
\item \code{sampledata} has to include columns that contain information on:
\itemize{
\item unique identifier of each sample; character or numeric (default column name = \code{Sample}),
\item date of sample collection in \code{Date} format (default = \code{Date}),
\item assignment of sample to particular individual; character or numeric (default = \code{AnimalRef}),
\item sex of the animal coded as \code{F}, \code{M} or \code{NA}; character (default = \code{GeneticSex}),
\item value that shows which sample is reference sample for each individual; boolean
coded as \code{1} or \code{0} (default = \code{IsAnimalReference}),
\item longitude and latitude coordinates of sample collection location; numeric (default = \code{X} and \code{Y}),
\item type of particular sample eg. scat, tissue, saliva; character (default = \code{SType}),
\item date of first and last sample of individual in \code{Date} format
(default = \code{FirstSeen} and \code{LastSeen}, see \code{\link{anim_timespan}} function),
\item value identifying if if the individual is dead; logical
(default = \code{IsDead}, see \code{\link{anim_timespan}} function).
}
}
}
\examples{
animal_ts <- anim_timespan(pack21_samples$AnimalRef,
                                  pack21_samples$Date,
                                  pack21_samples$SType,
                                  dead = c("Tissue", "Decomposing Tissue", "Blood"))

sampledata <- merge(pack21_samples, animal_ts, by.x = "AnimalRef", by.y = "ID", all.x = TRUE )

path <- paste0(system.file("extdata", package = "wpeR"), "/fake_colony")

ped_colony <- get_colony(path, sampledata, remove_obsolete_parents = TRUE, out = "FamAgg")

org_tables <- org_fams(ped_colony, sampledata, output = "both")


fam_table(org_tables$fams,
          org_tables$fams,
          org_tables$ped,
          sampledata,
          deadSample = c("Tissue", "Decomposing Tissue", "Blood"))

}
