% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_fams.R
\name{org_fams}
\alias{org_fams}
\alias{organizePacks}
\title{Organizes animals into families and expands pedigree data}
\usage{
org_fams(ped, sampledata, output = "both")
}
\arguments{
\item{ped}{data frame. \code{FamAgg} output of \code{\link{get_colony}} function.
With \code{RemoveObsoleteParents} parameter set to \code{TRUE}.}

\item{sampledata}{data frame. Metadata for all genetic samples that belong
to the individuals included in pedigree reconstruction analysis.
Must have $Sample with sample names and $GeneticSex coded as M/F/NA}

\item{output}{string. How the output data frame should be formatted.
output = "ped" returns pedigree data (similar to \code{\link{org_fams}} output)
with additional information on individuals. output = "fams" returns a table of all
families present in the pedigree. output = "both" returns a list with two
data frames: fams and ped. Defaults to "both"}
}
\value{
Based on the ´output´ parameter the function can return a data frame (\code{ped} or \code{fams})
or a list with two objects (\code{ped} and \code{fams}).
\itemize{
\item \code{ped} data frame. Extended output of \code{\link{org_fams}} function.
Apart from common pedigree information (individual, mother, father, sex, family), \code{ped}
also includes columns:
\itemize{
\item \code{parents}: identifier codes of both parents separated with \verb{_},
\item \code{FamID}: number of family that the individual belongs to (see \code{fams} below),
\item \code{FirstSeen}: date of first sample of individual,
\item \code{LastSeen}: date of last sample of individual,
\item \code{IsDead}: logical value (\code{TRUE/FALSE}) that identifies if the individual is dead,
\item \code{DadPclust}: identifier of fathers polygamy cluster,
\item \code{MomPclust}: identifier of mothers polygamy cuter,
\item \code{polyCluster}: polygamy cluster of the individual.
}
\item \code{fams} data frame includes information on families that individuals in the pedigree
belong to. The families are described by:
\itemize{
\item \code{parents}: identifier codes of both parents separated with \verb{_},
\item \code{father}: identifier code of the father,
\item \code{mother}: identifier code of the mother,
\item \code{FamID}: numeric value that identifies a particular family,
\item \code{famStart}: date when the first sample of any of the family members was collected,
\item \code{famEnd}: date when the last sample of any of the family members was collected,
\item \code{FamDead}: logical value (\code{TRUE/FALSE}) that identifies if the family does not exist any more,
\item \code{DadPclust}: identifier that connects families that share the same father,
\item \code{MomPclust}: identifier that connects families that share the same mother,
\item \code{polyCluster}: numeric value that connects families that share one of the parents.
}
}
}
\description{
\code{org_fams} takes pedigree data from \code{\link{get_colony}} function and groups animals into families.
It also expands the pedigree data by adding information about the family that each individual was born in and the
in which the individual is the reproductive animal. A family in this function is defined as a group of animals
where at least one parent and at least one offspring is known.
}
\examples{
animal_ts <- anim_timespan(pack21_samples$AnimalRef,
                                  pack21_samples$Date,
                                  pack21_samples$SType,
                                  dead = c("Tissue", "Decomposing Tissue", "Blood"))

sampledata <- merge(pack21_samples, animal_ts, by.x = "AnimalRef", by.y = "ID", all.x = TRUE )

path <- paste0(system.file("extdata", package = "wpeR"), "/fake_colony")

ped_colony <- get_colony(path, sampledata, remove_obsolete_parents = TRUE, out = "FamAgg")

org_fams(ped_colony, sampledata)

}
