% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pack21_samples}
\alias{pack21_samples}
\title{Wolf Pack Genetic Samples Data}
\format{
##\code{pack21_samples}
A data frame with 407 rows and 8 columns:
\describe{
\item{Sample}{Sample unique identifier code}
\item{Date}{Date of sample collection (format: \code{YYYY-MM-DD})}
\item{AnimalRef}{Identification string for patricular animal}
\item{GeneticSex}{Sex of animal to which the sample belong (format: \code{M} = male, \code{F} = female)}
\item{IsAnimalReference}{Is a samle also a reference sample for particular animal (format: \code{1} = yes, \code{0} = no)}
\item{X}{X coordinate (latitude) of the sample (CRS: WGS84; EPSG: 4326)}
\item{Y}{Y coordinate (longitude) of the sample (CRS: WGS84; EPSG: 4326)}
\item{SType}{Type of the sample. (\verb{Direct Saliva, Scat, Urine, Saliva, Tissue, Decomposing Tissue, Blood}) }
}
}
\source{
Slovenian National Wolf Monitoring
}
\usage{
pack21_samples
}
\description{
Data on samples of one wolf pack and its ancestor packs,
collected between 2010 and 2021, in the scope of Slovenian National Wolf Monitoring.
}
\keyword{datasets}
