% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_satplot.R
\name{ped_satplot}
\alias{ped_satplot}
\title{Mark-Recapture Plot For Pedigree - Saturating}
\usage{
ped_satplot(
  date,
  animal,
  plottingID,
  sex,
  fam,
  polyCluster,
  isPolygamous,
  rep,
  later_rep,
  dead,
  famSpacing = 2,
  pClustSpacing = 2,
  xWhiteSpace = 100,
  xlabel = "Date",
  ylabel = "Animal",
  title = "Pedigree CMR Graph",
  subtitle = "",
  LegendLabel = "Sex",
  xlegend = 0.2,
  ylegend = 0.94,
  text_size = 2.5,
  fam_label_size = 2
)
}
\arguments{
\item{date}{sample collection date. Must be in \code{Date} format.}

\item{animal}{individual animal identifier code.}

\item{plottingID}{ID for plotting, because of polygamy same animal can be in more families.}

\item{sex}{sex, but can be other parameters - you can set up LegendLabel}

\item{fam}{family identifier number.}

\item{polyCluster}{cluster for polygamous animals, includes all families in
which an individual is reproductive animal}

\item{isPolygamous}{logical value (\code{TRUE}/\code{FALSE}) that identifies
animal as polygamous (has >1 mates).}

\item{rep}{logical value (\code{TRUE}/\code{FALSE}) that determines if individual
is reproductive or not in current family.}

\item{later_rep}{logical value (\code{TRUE}/\code{FALSE}) that determines if individual
is reproductive or not in any other (later) family.}

\item{dead}{logical value (\code{TRUE/FALSE}) that identifies if the individual is dead}

\item{famSpacing}{Y-axis spacing between families. Shuld be even number!}

\item{pClustSpacing}{Y-axis spacing between families. Shuld be even number!}

\item{xWhiteSpace}{fill in}

\item{xlabel}{x axis label.}

\item{ylabel}{y axis label.}

\item{title}{plot title.}

\item{subtitle}{plot subtitle.}

\item{LegendLabel}{title of the legend.}

\item{xlegend}{horizontal position of the legend.}

\item{ylegend}{vertical position of the legend.}

\item{text_size}{plot text size.}

\item{fam_label_size}{family label text size.}
}
\value{
a graphical representation of detected family members trough time.
}
\description{
\code{ped_satplot} creates "capture" history plot of individuals
arranged by families included in data frame created by \code{\link{fam_table}} function.
}
\examples{
animal_ts <- anim_timespan(pack21_samples$AnimalRef,
                                  pack21_samples$Date,
                                  pack21_samples$SType,
                                  dead = c("Tissue", "Decomposing Tissue", "Blood"))

sampledata <- merge(pack21_samples, animal_ts, by.x = "AnimalRef", by.y = "ID", all.x = TRUE )

path <- paste0(system.file("extdata", package = "wpeR"), "/fake_colony")

ped_colony <- get_colony(path, sampledata, remove_obsolete_parents = TRUE, out = "FamAgg")

org_tables <- org_fams(ped_colony, sampledata, output = "both")

pt<-fam_table(org_tables$fams[1,],
              org_tables$fams,
              org_tables$ped,
              sampledata,
              deadSample = c("Tissue", "Decomposing Tissue", "Blood"))


ped_satplot(pt$Date,
                   pt$AnimalRef,
                   pt$plottingID,
                   pt$GeneticSex,
                   pt$FamID,
                   pt$polyCluster,
                   pt$isPolygamous,
                   pt$rep,
                   pt$later_rep,
                   pt$dead)







}
