% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_spatial.R
\name{ped_spatial}
\alias{ped_spatial}
\alias{PreparePedigreeSpatial}
\title{Get Files For Spatial Representation Of Pedigree}
\usage{
ped_spatial(
  plottable,
  na.rm = TRUE,
  output = "list",
  fullsibdata = NULL,
  sibthreshold = 0,
  path = "",
  filename = "",
  out.format = "geopackage",
  time.limits = c(as.Date("1900-01-01"), as.Date("2100-01-01")),
  time.limit.rep = FALSE,
  time.limit.offspring = FALSE,
  time.limit.moves = FALSE
)
}
\arguments{
\item{plottable}{data frame. Output of \code{\link{plot_table}} function.}

\item{na.rm}{logical (\code{TRUE}/\code{FALSE}). Remove samples with missing coordinates and/or dates.}

\item{output}{single value or vector. Type of output of function results.
Available outputs: list: all spatial data returned as list, gis: all spatial data
returned as georeferenced files.}

\item{fullsibdata}{data frame. COLONY2 fullsib data.}

\item{sibthreshold}{numeric. p value threshold for sibsihp assignment.}

\item{path}{string. System path where georeferenced files should be stored.}

\item{filename}{string. Common name for all georeferenced files.}

\item{out.format}{string. Type of georeferenced files to be generated. Can be eather
\code{"geopackage"} or \code{"shapefile"}. Default is \code{"geopackage"}}

\item{time.limits}{vector of two \code{Date} values. Time window for movement and
offspring reference samples data.}

\item{time.limit.rep}{logical (\code{TRUE}/\code{FALSE}). Do time limits apply to
reference samples of reproductive animals?}

\item{time.limit.offspring}{logical (\code{TRUE}/\code{FALSE}). Do time limits also apply
to reference samples of offspring?}

\item{time.limit.moves}{logical (\code{TRUE}/\code{FALSE}). Do time limits also apply
to movement data?}
}
\value{
Based on the \code{output} parameter the function can return a list of \code{\link{sf}} objects,
a geospatial vector data files or both.

Most of the objects are created separately for mothers, fathers and offspring,
this include: reference points (\code{motherRpoints}, \code{fatherRpoints} and
\code{offspringRpoints}), movement points (\code{motherMovePoints}, \code{fatherMovePoints}
and \code{offspringMovePoints}), movement lines (\code{motherMoveLines}, \code{fatherMoveLines}
and \code{offspringMoveLines}) and movement polygons (\code{motherMovePoygons},
\code{fatherMovePolygons} and \code{offspringMovePolygons}).

Besides that the function also produces lines that connect mothers and
their offspring (\code{maternityLines}), fathers and their offspring
(\code{paternityLines}) and full siblings (\code{FullsibLines}).
}
\description{
\code{ped_spatial} creates georeferenced data for spatial
pedigree representation form the output of \code{\link{plot_table}} function.
}
\examples{
animal_ts <- anim_timespan(wolf_samples$AnimalRef,
                                  wolf_samples$Date,
                                  wolf_samples$SType,
                                  dead = c("Tissue", "Decomposing Tissue", "Blood"))

sampledata <- merge(wolf_samples, animal_ts, by.x = "AnimalRef", by.y = "ID", all.x = TRUE )

path <- paste0(system.file("extdata", package = "wpeR"), "/wpeR_samplePed")

ped_colony <- get_colony(path, sampledata, remove_obsolete_parents = TRUE, out = "FamAgg")

org_tables <- org_fams(ped_colony, sampledata, output = "both")

pt<-plot_table(org_tables$fams[1,],
              org_tables$fams,
              org_tables$ped,
              sampledata,
              deadSample = c("Tissue", "Decomposing Tissue", "Blood"))


ped_spatial(pt)


}
