% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Prepares pedigree data for plotting and spatial representation}
\usage{
plot_table(
  plot_fams = NULL,
  plot_indivs = NULL,
  all_fams,
  ped,
  sampledata,
  datacolumns = c("Sample", "AnimalRef", "GeneticSex", "Date", "SType", "lat", "lng",
    "FirstSeen", "LastSeen", "IsDead", "IsMortality")
)
}
\arguments{
\item{plot_fams}{Character string or numeric vector. Defines which families to include.
Use "all" (or NULL, default) to include all families, or a numeric vector of specific \code{FamID}
numbers to plot a subset (see Details).}

\item{plot_indivs}{Character string or vector of character strings. Individual identifiers to include.
The families associated with these individuals (as father, mother, or offspring)
will be included in the output (see Details).}

\item{all_fams}{Data frame. Family (\code{fams}) data generated by \code{\link[=org_fams]{org_fams()}} function.}

\item{ped}{Data frame. Organized pedigree (\code{ped}) generated by \code{\link[=org_fams]{org_fams()}} function.}

\item{sampledata}{Data frame. Metadata for all genetic samples that belong
to the individuals included in pedigree reconstruction analysis.
For description of \code{sampledata} structure and sample information needed for \code{plot_table()} see Details.}

\item{datacolumns}{Vector of column names included \code{sampledata} that are needed to produce
this functions output (see Details).}
}
\value{
Extended \code{sampledata} data frame that includes all columns defined in \code{datacolumns}
parameter and adds information needed for visual and spatial representation of
pedigree:
\itemize{
\item \code{plottingID}: Numeric. Identifier number for temporal pedigree plot
\code{\link[=ped_satplot]{ped_satplot()}}. In case of polygamous animals same individual
can be included in more than one family.
\item \code{FamID}: Numeric. Identifier number of family that individual belongs to.
\item \code{hsGroup}: Numeric. Identifier number for the half-sib group of individual.
\item \code{rep}: Logical. Is individual reproductive in current family,
(current family defined with \code{FamID} for a particular entry).
\item \code{later_rep}: Logical. Is individual reproductive in any other (later) families.
\item \code{isPolygamous}: Logical. Does the individual have more than one mate.
\item \code{dead}: Logical. Is individual dead.
\item \code{first_sample}: Logical. Is this particular sample the first sample of the individual.
\item \code{last_sample}: Logical. Is this particular sample the last sample of the individual.
}
}
\description{
Combines extended pedigree (obtained by \code{\link[=org_fams]{org_fams()}} function)
and sample metadata data for visual (\code{\link[=ped_satplot]{ped_satplot()}}) and spatial
(\code{\link[=ped_spatial]{ped_spatial()}}) representation of the pedigree.
}
\details{
\itemize{
\item \code{sampledata} has to include columns that contain information on:
\itemize{
\item unique identifier of each sample; character or numeric
(default column name = \code{Sample}, see \code{\link[=check_sampledata]{check_sampledata()}} function),
\item date of sample collection in \code{Date} format (default = \code{Date}),
\item assignment of sample to particular individual; character or numeric
(default = \code{AnimalRef}, see \code{\link[=check_sampledata]{check_sampledata()}} function),
\item sex of the animal coded as \code{F}, \code{M} or \code{NA}; character
(default = \code{GeneticSex}, see \code{\link[=check_sampledata]{check_sampledata()}} function),
\item longitude and latitude coordinates of sample collection location; numeric
(default = \code{lng} and \code{lat}, see \code{\link[=check_sampledata]{check_sampledata()}} function),
\item type of particular sample eg. scat, tissue, saliva; character
(default = \code{SType}, see \code{\link[=check_sampledata]{check_sampledata()}} function),
\item date of first and last sample of individual in \code{Date} format
(default = \code{FirstSeen} and \code{LastSeen}, see \code{\link[=anim_timespan]{anim_timespan()}} function),
\item value identifying if the individual is dead; logical
(default = \code{IsDead}, see \code{\link[=anim_timespan]{anim_timespan()}} function).
\item value identifying if this particular sample represents a mortality sample
(default = \code{IsMortality}, see \code{\link[=check_sampledata]{check_sampledata()}} function)
}
}

\strong{Filtering Options}
The \code{plot_table} function allows for flexible subsetting of families to be plotted
using the \code{plot_fams} and \code{plot_indivs} arguments:
\itemize{
\item \code{plot_fams}: A numeric vector of specific \code{FamID}s to include.
\item \code{plot_indivs}: A vector of individual identifiers. Families where these individuals
appear as a father, mother, or offspring are automatically identified and included.
}

Behavior:
\itemize{
\item If both \code{plot_fams} and \code{plot_indivs} are \code{NULL} (default), all families are included.
\item If one or both are provided, the final set of families is the \strong{union} of
families identified by both filters.
}
}
\examples{

# Prepare the data for usage with plot_table() function.
# Get animal timespan data using the anim_timespan() function.
animal_ts <- anim_timespan(wolf_samples$AnimalRef,
  wolf_samples$Date,
  wolf_samples$IsMortality
)
# Add animal timespan to the sampledata
sampledata <- merge(wolf_samples, animal_ts, by.x = "AnimalRef", by.y = "ID", all.x = TRUE)
# Define the path to the pedigree data file.
path <- paste0(system.file("extdata", package = "wpeR"), "/wpeR_samplePed")
# Retrieve the pedigree data from the get_colony function.
ped_colony <- get_colony(path, sampledata, rm_obsolete_parents = TRUE, out = "FamAgg")
# Organize families and expand pedigree data using the org_fams function.
org_tables <- org_fams(ped_colony, sampledata, output = "both")

# Run the function
# Standard usage: all families
pt <- plot_table(plot_fams = "all",
  all_fams = org_tables$fams,
  ped = org_tables$ped,
  sampledata = sampledata,
)


# Plot table for a subset of individuals
pt_subset <- plot_table(plot_indivs = "M200F",
                        all_fams = org_tables$fams,
                        ped = org_tables$ped,
                        sampledata = sampledata
                       )

}
