% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolf_samples_old}
\alias{wolf_samples_old}
\title{Wolf Monitoring Genetic Samples Data}
\format{
A data frame with 407 rows and 8 columns:
\describe{
\item{Sample}{Sample unique identifier code}
\item{Date}{Date of sample collection (format: \code{YYYY-MM-DD})}
\item{AnimalRef}{Identification string for patricular animal}
\item{GeneticSex}{Sex of animal to which the sample belong (format: \code{M} = male, \code{F} = female)}
\item{IsAnimalReference}{Is a samle also a reference sample for particular animal (format: \code{1} = yes, \code{0} = no)}
\item{lat}{latitude (N-S) of the sample (CRS: WGS84; EPSG: 4326)}
\item{lng}{longitude (W-E) of the sample (CRS: WGS84; EPSG: 4326)}
\item{SType}{Type of the sample. (\verb{Direct Saliva, Scat, Urine, Saliva, Tissue, Decomposing Tissue, Blood}) }
}
}
\source{
Slovenian National Wolf Monitoring
}
\usage{
wolf_samples_old
}
\description{
Metadata of selected genetic samples of wolves collected between XXXX and YYYY,
in the scope of Slovenian National Wolf Monitoring
}
\keyword{datasets}
