% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sampledata.R
\name{check_sampledata}
\alias{check_sampledata}
\title{Check and prepare sample data}
\usage{
check_sampledata(
  Sample,
  Date,
  AnimalRef,
  GeneticSex,
  IsAnimalReference,
  lat,
  lng,
  SType
)
}
\arguments{
\item{Sample}{A vector of sample unique identifier codes.}

\item{Date}{A vector of sample collection dates in 'YYYY-MM-DD' format.}

\item{AnimalRef}{A vector identifier codes of the particular individual that the sample belongs to.}

\item{GeneticSex}{A vector of genetic sex information ('F' for female, 'M' for male, NA for unknown).}

\item{IsAnimalReference}{A vector indicating whether the sample is an animal reference (0 or 1).}

\item{lat}{A vector of latitude coordinates in the WGS84 coordinate system.}

\item{lng}{A vector of longitude coordinates in the WGS84 coordinate system.}

\item{SType}{A vector of sample types.}
}
\value{
A well-structured sample data frame with validated and formatted columns.
}
\description{
This function checks the integrity and compatibility of various columns in the sample data
and prepares a well-structured sample data frame. It verifies the consistency of columns
such as Sample, Date, AnimalRef, GeneticSex, IsAnimalReference, lat, lng, and SType.
The function ensures that the provided data is properly formatted and conforms to the standards
of functions that make up the \code{wpeR} package.
}
\examples{
sampledata <- check_sampledata(
  Sample = wolf_samples$Sample,
  Date = wolf_samples$Date,
  AnimalRef = wolf_samples$AnimalRef,
  GeneticSex = wolf_samples$GeneticSex,
  IsAnimalReference = wolf_samples$IsAnimalReference,
  lat = wolf_samples$lat,
  lng = wolf_samples$lng,
  SType = wolf_samples$SType
)

}
