% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sampledata.R
\name{check_sampledata}
\alias{check_sampledata}
\title{Check and Prepare Sample Data}
\usage{
check_sampledata(
  Sample,
  Date,
  AnimalRef,
  GeneticSex,
  IsAnimalReference,
  lat,
  lng,
  SType
)
}
\arguments{
\item{Sample}{A vector of sample identifiers.}

\item{Date}{A vector of sample dates in 'YYYY-MM-DD' format.}

\item{AnimalRef}{A vector of animal reference identifiers.}

\item{GeneticSex}{A vector of genetic sex information ('F' for female, 'M' for male, NA for unknown).}

\item{IsAnimalReference}{A vector indicating whether the sample is an animal reference (0 or 1).}

\item{lat}{A vector of latitude coordinates in the WGS84 coordinate system.}

\item{lng}{A vector of longitude coordinates in the WGS84 coordinate system.}

\item{SType}{A vector of sample types.}
}
\value{
A well-structured sample data frame with validated and formatted columns.
}
\description{
This function checks the integrity and compatibility of various columns in the sample data
and prepares a well-structured sample data frame. It verifies the consistency of columns
such as Sample, Date, AnimalRef, GeneticSex, IsAnimalReference, lat, lng, and SType.
The function ensures that the provided data is properly formatted and conforms to the required standards.
}
\examples{
sampledata <- check_sampledata(Sample = c("S1", "S2"),
                               Date = c("2023-01-15", "2023-02-20"),
                               AnimalRef = c("A1", "A2"),
                               GeneticSex = c("F", "M"),
                               IsAnimalReference = c(1, 0),
                               lat = c(34.0522, 40.7128),
                               lng = c(-118.2437, -74.0060),
                               SType = c("Type1", "Type2"))

}
