% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ped.R
\name{get_ped}
\alias{get_ped}
\title{Organizes Pedigree Data}
\usage{
get_ped(ped, sampledata, out = "FamAgg")
}
\arguments{
\item{ped}{Data frame. Pedigree data frame with the most basic structure.
Four columns corresponding to offspring, father, mother and cluster (see
Details). Unknown parents should be represented by \code{NA} values.}

\item{sampledata}{Data frame. Metadata for all genetic samples that belong
to the individuals included in pedigree reconstruction analysis.
This data frame should adhere to the formatting and naming conventions
outlined in the \code{\link[=check_sampledata]{check_sampledata()}} documentation.}

\item{out}{Character string. For use with which package should the output be formatted?
\code{kinship2} (out = "kinship2"), \code{pedtools} (out = "pedtools") or
\code{FamAgg} (out = "FamAgg"). Defaults to "FamAgg"}
}
\value{
A data frame describing a common pedigree structure. Each individual included in
pedigree represents one row. Columns describe individual identifier code, identifier code for
mother and father and sex of individual. Column names and arrangement depends on selected
output (\code{out} parameter).
}
\description{
Offers an alternative to \code{\link[=get_colony]{get_colony()}} function in cases where the pedigree
was not reconstructed with \href{https://www.zsl.org/about-zsl/resources/software/colony}{COLONY 2}
software. It takes a pedigree dataframe and assigns sex to each individual.
The function also prepares data so that the output of the function can be directly analysed with
\href{https://cran.r-project.org/web/packages/kinship2/index.html}{\code{kinship2}},
\href{https://cran.r-project.org/web/packages/pedtools/index.html}{\code{pedtools}} or
\href{https://bioconductor.org/packages/release/bioc/html/FamAgg.html}{\code{FamAgg}} packages.
}
\details{
The custom pedigree specified through the \code{ped} parameter should mirror the
structure of a COLONY2 pedigree and share the same column names.
It should consist of four columns for each offspring:
\code{OffspringID}, \code{FatherID}, \code{MotherID} and \code{ClusterIndex}. In the context of
COLONY2 result, the \code{ClusterIndex} refers to a group of offspring that may
share common parents or ancestors and are analyzed together. If your
your pedigree does not include such information you can fill this column with
the same numeric value (eg. 1, see examples) or any (numeric) information
about other family structure present in your pedigree. When considering
unknown parents they should be represented by \code{NA} values.
}
\examples{
#example pedigree dataframe
ped <- data.frame(
  OffspringID = c(
    "M273P", "M20AM", "M2757", "M2ALK", "M2ETE", "M2EUJ", "MSV00E",
    "MSV018", "MSV05L", "MSV0M6", "MSV0T4", "MSV0T7", "MSV0TJ", "MSV0UL"
  ),
  FatherID = c(
    NA, NA, "M20AM", "M20AM", "M20AM", "M20AM", "M20AM",
    "M20AM", "M20AM", "M20AM", "M20AM", "M20AM", "M20AM", "M20AM"
  ),
  MotherID = c(
    NA, NA, "M273P", "M273P", "M273P", "M273P", "M273P",
    "M273P", "M273P", "M273P", "M273P", "M273P", "M273P", "M273P"
  ),
  ClusterIndex = c(rep(1, 14))
)
#Get pedigree data in FamAgg format
get_ped(
    ped = ped,
    sampledata = wolf_samples
    )

}
