% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolf_samples}
\alias{wolf_samples}
\title{Wolf monitoring genetic samples metadata}
\format{
A data frame with 407 rows and 7 columns:
\describe{
\item{Sample}{Sample unique identifier code}
\item{Date}{Date of sample collection (format: \code{YYYY-MM-DD})}
\item{AnimalRef}{Identification string for particular animal}
\item{GeneticSex}{Sex of animal to which the sample belong (format: \code{M} = male, \code{F} = female)}
\item{lat}{latitude (N-S) of the sample (CRS: WGS84; EPSG: 4326)}
\item{lng}{longitude (W-E) of the sample (CRS: WGS84; EPSG: 4326)}
\item{SType}{Type of the sample. (\verb{Direct Saliva, Scat, Urine, Saliva, Tissue, Decomposing Tissue, Blood}) }
}
}
\source{
Slovenian National Wolf Monitoring
}
\usage{
wolf_samples
}
\description{
Metadata of selected genetic samples of wolves collected between 2015 and 2021,
in the scope of Slovenian National Wolf Monitoring
}
\keyword{datasets}
